+++ config/test.exs
@@ -16,8 +16,7 @@ config :brando, Brando.Integration.TestRepo,
   max_overflow: 0
 
 config :brando, Brando.Menu, [
-  modules: [Brando.Menu.Admin, Brando.Menu.Users,
-            Brando.Menu.News, Brando.Menu.Images],
+  modules: [Brando.Admin.Menu, Brando.Users.Menu, Brando.Images.Menu],
   colors: ["#FBA026;", "#F87117;", "#CF3510;", "#890606;", "#FF1B79;",
            "#520E24;", "#8F2041;", "#DC554F;", "#FF905E;", "#FAC51C;",
            "#D6145F;", "#AA0D43;", "#7A0623;", "#430202;", "#500422;",